locust/requirements.txt
locust==2.15.1
psycopg2-binary==2.9.6